refactor: decompose ScholarshipForm into step components
Extract GeneralInfoStep and EligibilityStep from ScholarshipForm.tsx
into src/components/scholarship-form/. The main ScholarshipForm now
delegates step content to these components.

- GeneralInfoStep: name, organization, website, deadline, amount, description, tags
- EligibilityStep: lint warnings, no-reqs checkbox, grades, GPA, schools, states, majors, ethnicity
- ScholarshipForm: stepper, validation, formik, onSubmit, autoFill orchestration
